+++ controllers/ResultController.php
@@ -35,6 +35,15 @@ class ResultController extends Controller
      */
     public function actionIndex()
     {
+
+        if (Yii::$app->user->isGuest) {
+
+            Yii::$app->session->setFlash('error', "You are not log in!");
+            return $this->redirect('http://app.test/site/login');
+
+
+        }
+
         $searchModel = new ResultSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -52,6 +61,14 @@ class ResultController extends Controller
      */
     public function actionView($id)
     {
+        if (Yii::$app->user->isGuest) {
+
+            Yii::$app->session->setFlash('error', "You are not log in!");
+            return $this->redirect('http://app.test/site/login');
+
+
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -64,6 +81,14 @@ class ResultController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->isGuest) {
+
+            Yii::$app->session->setFlash('error', "You are not log in!");
+            return $this->redirect('http://app.test/site/login');
+
+
+        }
+
         $model = new Result();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -84,6 +109,14 @@ class ResultController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->isGuest) {
+
+            Yii::$app->session->setFlash('error', "You are not log in!");
+            return $this->redirect('http://app.test/site/login');
+
+
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -104,6 +137,13 @@ class ResultController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->isGuest) {
+
+            Yii::$app->session->setFlash('error', "You are not log in!");
+            return $this->redirect('http://app.test/site/login');
+
+
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
